scripts/content-automation/requirements.txt
openai>=1.0.0
anthropic>=0.18.0
supabase>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
markdown>=3.5.0

